[fix] Moved delta extension into a single module
