+++ aptos-move/aptos-vm/src/delta_ext/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Aptos
+// SPDX-License-Identifier: Apache-2.0
+
+mod delta_change_set;
+mod transaction;
+
+pub use crate::delta_ext::{
+    delta_change_set::{deserialize, serialize, DeltaChangeSet, DeltaOp},
+    transaction::{ChangeSetExt, TransactionOutputExt},
+};
