+++ types/Cargo.toml
@@ -35,7 +35,6 @@ aptos-crypto-derive = { path = "../crates/aptos-crypto-derive" }
 move-deps = { path = "../aptos-move/move-deps", features = ["address32"] }
 
 [dev-dependencies]
-claim = "0.5.0"
 proptest = "1.0.0"
 proptest-derive = "0.3.0"
 regex = "1.5.5"
