+++ types/src/lib.rs
@@ -11,7 +11,6 @@ pub mod block_info;
 pub mod block_metadata;
 pub mod chain_id;
 pub mod contract_event;
-pub mod delta_change_set;
 pub mod epoch_change;
 pub mod epoch_state;
 pub mod event;
